Partner_OS_V1/requirements.txt
streamlit>=1.42,<2
requests>=2.32,<3
python-dotenv>=1.0,<2
pydantic>=2.10,<3
pytest>=8.3,<9
